git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/branches/issue-CLO-126@13257 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.engine/src/org/jetel/metadata/DataRecordMetadata.java
@@ -1308,7 +1308,7 @@ public class DataRecordMetadata implements Serializable, Iterable<DataFieldMetad
 	public String getFieldNamesHeader(String[] excludedFieldNames, boolean quotedString, Character quoteChar) {
 		StringBuilder ret = new StringBuilder();
 		DataFieldMetadata dataFieldMetadata;
-		short fieldSize;
+		int fieldSize;
 		int headerSize;
 		char blank = ' ';
 		String label;
